webapp/requirements.txt
Django==4.1.10
djangorestframework>=3.12.4,<3.13
psycopg2-binary
django-cors-headers>=3.10.0,<3.11
celery
numpy
pandas
scikit-learn
joblib